ModoInfinito pt.2
Cambié valores de int a float. Agregué NexusManager y ScoreManager para que los enemigos agreguen puntaje al ser eliminados y dañen el nexo al llegar a el.

+++ Assets/Scripts/Players/Bala.cs
@@ -8,7 +8,7 @@ public class Bala : MonoBehaviour
 {
     [SerializeField] public Rigidbody rigidbodyBala;
     [SerializeField] public float velocidad;
-    [SerializeField] public int dañoBala = 1;
+    [SerializeField] public float dañoBala = 1;
 
     private void Awake()
     {
